corrected and added code
